+++ Example/DKHelperTests/NSDateTests.swift
@@ -106,14 +106,21 @@ extension NSDateTests {
 extension NSDateTests {
 
 	func test_ShouldReturnValidAndCorrectDate() {
-		let date = NSDate(fromString: "24.12.1992", style: .MediumStyle)
+		var string = "Dec 6, 1992, 2:59:10 PM" // en_US
+		if (NSLocale.currentLocale().localeIdentifier == "en_DE") {
+			string = "06 Dec 1992 14:35:32"
+		} else if (NSLocale.currentLocale().localeIdentifier == "de_DE") {
+			string = "06.12.1992, 14:35:32"
+		}
+
+		let date = NSDate(fromString: string, style: .MediumStyle)
 		XCTAssertNotNil(date)
 		XCTAssertEqual(date?.year(), 1992)
-		XCTAssertEqual(date?.day(), 24)
+		XCTAssertEqual(date?.day(), 6)
 		XCTAssertEqual(date?.month(), 12)
 	}
 
-	func test_ShouldReturnInValidDate() {
+	func test_ShouldReturnInvalidDate() {
 		let date = NSDate(fromString: "Dec 24, 1992", style: .MediumStyle)
 		let dateByFormat = NSDate(fromString: "24.12.1992", format: "yyyy.MM.dd")
 		let dateByString = NSDate(fromString: "ab.cd.efgh", format: "dd.MM.yyyy")
@@ -209,12 +216,17 @@ extension NSDateTests {
 		XCTAssertNotNil(date)
 		XCTAssertNotNil(dayName)
 		XCTAssertNotNil(monthName)
-		XCTAssertEqual(dayName, "Freitag", "expect german localization")
-		XCTAssertEqual(monthName, "Dezember", "expect german localization")
+		if (NSLocale.currentLocale().localeIdentifier.hasPrefix("en") == true) {
+			XCTAssertEqual(dayName, "Friday", "expect english localization")
+			XCTAssertEqual(monthName, "December", "expect english localization")
+		} else if (NSLocale.currentLocale().localeIdentifier.hasPrefix("de") == true) {
+			XCTAssertEqual(dayName, "Freitag", "expect german localization")
+			XCTAssertEqual(monthName, "Dezember", "expect german localization")
+		}
 	}
 
 	func test_ShouldReturnValidAndCorrectDateDisplayStrings() {
-		let date = NSDate(fromString: "04.12.2015 - 14:27:11", format: "dd.MM.yyyy - HH:mm:ss")
+		let date = NSDate(fromString: "04.12.2015 - 15:27:11", format: "dd.MM.yyyy - HH:mm:ss")
 		let fullDisplayString = date?.fullDisplayTime()
 		let hourDisplayString = date?.hourDisplayTime()
 		let displayString = date?.displayableString()
@@ -222,9 +234,37 @@ extension NSDateTests {
 		XCTAssertNotNil(fullDisplayString)
 		XCTAssertNotNil(hourDisplayString)
 		XCTAssertNotNil(displayString)
-		XCTAssertEqual(fullDisplayString, "4. Dezember 2015 um 15:27", "expect german localization")
-		XCTAssertEqual(hourDisplayString, "15:27", "expect german localization")
-		XCTAssertEqual(displayString, "04.12.2015", "expect german localization")
+
+		if (NSLocale.currentLocale().localeIdentifier.hasPrefix("de") == true) {
+			XCTAssertEqual(fullDisplayString, "4. Dezember 2015 um 15:27", "expect german localization")
+			XCTAssertEqual(hourDisplayString, "15:27", "expect german localization")
+			XCTAssertEqual(displayString, "04.12.2015", "expect german localization")
+		} else if (NSLocale.currentLocale().localeIdentifier.hasPrefix("en_DE") == true) {
+			XCTAssertEqual(fullDisplayString, "4 December 2015 at 15:27", "expect english localization with german region")
+			XCTAssertEqual(hourDisplayString, "15:27", "expect english localization with german region")
+			XCTAssertEqual(displayString, "04 Dec 2015", "expect english localization with german region")
+		} else if (NSLocale.currentLocale().localeIdentifier.hasPrefix("en") == true) {
+			XCTAssertEqual(fullDisplayString, "December 4, 2015 at 3:27 PM", "expect english localization")
+			XCTAssertEqual(hourDisplayString, "3:27 PM", "expect english localization")
+			XCTAssertEqual(displayString, "Dec 4, 2015", "expect english localization")
+		}
+	}
+}
+
+// MARK: - Log methods
+
+extension NSDateTests {
+
+	func test_ShouldLogAllFormat() {
+		let date = NSDate()
+		date.logAllFormats()
+		XCTAssert(true)
+	}
+
+	func test_ShouldLogOneFormat() {
+		let date = NSDate()
+		date.logCurrentDateWithDateStyleAndAllTimeStyle(.FullStyle)
+		XCTAssert(true)
 	}
 }